pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "hippoclaudus"
version = "1.1.0"
description = "Local LLM-powered memory management for Claude's persistent memory system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.5",
    "mlx-lm>=0.10",
]
llama-cpp = [
    "llama-cpp-python>=0.2",
]

[tool.setuptools.packages.find]
include = ["hippoclaudus*"]
